+++ src/Skill/DependencyInjection/Container.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jarvis\DependencyInjection;
+namespace Jarvis\Skill\DependencyInjection;
 
 /**
  * Minimalist dependency injection container
