+++ src/Skill/EventBroadcaster/AnalyzeEvent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jarvis\EventBroadcaster;
+namespace Jarvis\Skill\EventBroadcaster;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
