+++ tests/JarvisBroadcasterTest.php
@@ -3,13 +3,13 @@
 namespace Jarvis\Tests;
 
 use Jarvis\Jarvis;
-use Jarvis\DependencyInjection\Reference;
-use Jarvis\EventBroadcaster\AnalyzeEvent;
-use Jarvis\EventBroadcaster\ControllerEvent;
-use Jarvis\EventBroadcaster\EventInterface;
-use Jarvis\EventBroadcaster\JarvisEvents;
-use Jarvis\EventBroadcaster\SimpleEvent;
-use Jarvis\EventBroadcaster\ResponseEvent;
+use Jarvis\Skill\DependencyInjection\Reference;
+use Jarvis\Skill\EventBroadcaster\AnalyzeEvent;
+use Jarvis\Skill\EventBroadcaster\ControllerEvent;
+use Jarvis\Skill\EventBroadcaster\EventInterface;
+use Jarvis\Skill\EventBroadcaster\JarvisEvents;
+use Jarvis\Skill\EventBroadcaster\SimpleEvent;
+use Jarvis\Skill\EventBroadcaster\ResponseEvent;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
